server/requirements.txt
screenpoint==0.1.2
Flask==1.1.1
flask-cors==3.0.9
pyscreenshot==2.2
Pillow==7.2.0
requests==2.24.0
opencv-python==4.4.0.44